feat(AIAI-575): statistic charts
* feat(AIAI-575): initialize smart-workflow-dashboard project

* fix(AIAI-575): use toString() for model name listing

* feat(AIAI-575): add governance dashboard backend

* feat(AIAI-575): add governance dashboard UI

* test(AIAI-575): add governance dashboard web tests

* fix(AIAI-575): move governance history classes out of internal packages

* fix(AIAI-575): fixed gui test

* fix(AIAI-575): added `test.engine.url` when run web test

* fix(AIAI-575): added `test.engine.url` when run web test

* fix(AIAI-575): added `test.engine.app` when run web test

* fix(AIAI-575): fix integration test

* fix(AIAI-575): run web test

* fix(AIAI-575): added test user

* fix(AIAI-575): build smart-workflow-dashboard when run web test

* fix(AIAI-575): correct test cases

* fix(AIAI-575): updated test cases

* fix(AIAI-575): move UI-only classes to dashboard, fix webtest setup

* fix(AIAI-575): added variables when run test

* fix(AIAI-575): try new way to add variables when run test

* fix(AIAI-575): try another way to add variables when run test

* fix(AIAI-575): try another way to add variables when run test

* fix(AIAI-575): try another way to add variables when run test

* feat(AIAI-575): add usage analytics chart builders and aggregator

* feat(AIAI-575): add analytics dashboard page with charts, CMS labels, and assets

* feat(AIAI-575): improve history JSON parsing and UI entity view models

* test(AIAI-575): add chart builder, history filter, and history tree unit tests

* fix(AIAI-575): run web-integration-test with engine props

* fix; run web-integration-test with engine props

* Try Chrome

* changed pom.xml

* fix: added back missing code after merge

* fix: updated version for build plugin

* updated ivy-project-parent version

---------


* test(AIAI-575): Web test for statistic charts

* fix; run web-integration-test with engine props

* Try Chrome

* changed pom.xml

* fix: added back missing code after merge

* fix: updated version for build plugin

* test(AIAI-575): updated web test for statistic feature

---------


* feat: upgrade project

* feat(AIAI-575): convert to latest ivy version

* fix: compile error after convert

---------

+++ smart-workflow-dashboard/src/com/axonivy/utils/smart/workflow/governance/ui/entity/AgentConversationView.java
@@ -3,11 +3,11 @@ package com.axonivy.utils.smart.workflow.governance.ui.entity;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import com.axonivy.utils.smart.workflow.governance.history.ChatHistoryJsonParser;
+import com.axonivy.utils.smart.workflow.governance.history.ChatHistoryJsonParser.TokenUsage;
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry;
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry.GuardrailExecution;
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry.ToolExecution;
-import com.axonivy.utils.smart.workflow.governance.history.ChatHistoryJsonParser;
-import com.axonivy.utils.smart.workflow.governance.history.ChatHistoryJsonParser.TokenUsage;
 import com.axonivy.utils.smart.workflow.governance.utils.DatePatternUtils;
 
 public class AgentConversationView {
@@ -45,7 +45,7 @@ public class AgentConversationView {
     return ChatHistoryJsonParser.getMessageCount(entry);
   }
 
-  public int getTotalTokens() {
+  public long getTotalTokens() {
     return tokenUsage().totalTokens();
   }
